backend/requirements.txt
flask==3.0.2
python-dotenv==1.0.1
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
openai==1.60.0 